Make email unique + error message
I checked that email would be unique in DB by making it unique in the module, I added a check in case of an error in writing to DB, for the error clearly I used external directory.
And more repairs ...
I added a router to the project for easy and neat navigation.
And I added the check whether the request is a POST type into the middleware that will always be called before the request.